requirements.txt
# nltk~=3.6.2
# numpy~=1.21.4
# tqdm~=4.61.0
# requests~=2.25.1
# matplotlib~=3.4.2
# torch~=1.10.0
# transformers~=4.12.5
# pandas~=1.2.4
prior
emissor
cltl.combot
ollama
jupyter
llama-index-core
llama-index-utils-workflow
openai==1.40.0
python-dotenv
ai2thor